Move staff routes under (data-view) layout group and refactor ServiceForm
- Replace `<a>` with `<Link>` in BrandButton for client-side navigation
- Add `prompt: "select_account"` to Google OAuth config
- Convert Edit button from `<button>` to accessible `<span>` with
  keyboard support
- Move Danger Zone into ServiceForm `footer` prop
- Clean up ServiceForm scroll effect and adopt DataViewPageHeader
