+++ convex/auth.ts
@@ -99,6 +99,7 @@ export const createAuthOptions = (ctx: GenericCtx<DataModel>) => {
           google: {
             clientId: process.env.GOOGLE_CLIENT_ID as string,
             clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+            prompt: "select_account",
           },
         },
