+++ src/app/(private)/dashboard/(staff)/(data-view)/services/[slug]/_client.tsx
