+++ src/app/(private)/dashboard/(staff)/(data-view)/services/add-service/page.tsx
@@ -1,0 +1,5 @@
+import { AddServiceClient } from "./_client";
+
+export default function AddServicePage() {
+  return <AddServiceClient />;
+}
