+++ src/app/(private)/dashboard/(staff)/(data-view)/workshops/[slug]/edit/_client.tsx
@@ -152,20 +152,18 @@ export function EditWorkshopClient({
   };
 
   return (
-    <>
-      <ServiceForm
-        title={`Edit ${service.name}`}
-        initialValues={initialValues}
-        initialImages={initialImages}
-        initialSamples={initialSamples}
-        onSubmit={handleSubmit}
-        onDiscard={() => router.push("/dashboard/workshops")}
-        submitError={submitError}
-        mode="WORKSHOP"
-        backHref="/dashboard/workshops"
-      />
-      <div className="container mx-auto max-w-6xl px-10 pb-10 -mt-8">
-        <div className="mt-12 rounded-lg border border-destructive bg-destructive/5 p-6 flex flex-col sm:flex-row items-start sm:items-center justify-between gap-4 mb-8">
+    <ServiceForm
+      title={`Edit ${service.name}`}
+      initialValues={initialValues}
+      initialImages={initialImages}
+      initialSamples={initialSamples}
+      onSubmit={handleSubmit}
+      onDiscard={() => router.push("/dashboard/workshops")}
+      submitError={submitError}
+      mode="WORKSHOP"
+      backHref="/dashboard/workshops"
+      footer={
+        <div className="rounded-lg border border-destructive bg-destructive/5 p-6 flex flex-col sm:flex-row items-start sm:items-center justify-between gap-4">
           <div className="space-y-1">
             <h3 className="font-semibold text-destructive">Danger Zone</h3>
             <p className="text-sm text-muted-foreground">
@@ -182,7 +180,7 @@ export function EditWorkshopClient({
             className="bg-destructive text-white hover:bg-destructive/90"
           />
         </div>
-      </div>
-    </>
+      }
+    />
   );
 }
