+++ src/app/(private)/dashboard/(staff)/(data-view)/workshops/[slug]/edit/page.tsx
