+++ src/app/(private)/dashboard/(staff)/(data-view)/workshops/add-workshop/page.tsx
@@ -1,0 +1,5 @@
+import { AddWorkshopClient } from "./_client";
+
+export default function AddWorkshopPage() {
+  return <AddWorkshopClient />;
+}
